Refactor character conversion logic
Specified ranges of character conversion to distinguish between upper
and lower case characters. Used the module operator to quantify the
remainder from a 26 division to keep character shifted by the cipher key
in their respective ranges.
The logic is that if a shifted character is still in range the remainder
is = 0, so no correction is needed, if it's not the remainder is > 0 and
it is equal to the amount of shifts needed from the start of the range.
Refactored the code to keep it short and concise.