Refactor the code to include sub-strings count
Removed select and added iteration with #each that pushes matching word
into the matching_words_arr.
To achieve sub-string matching a second iteration using #each goes
through the given collection and checks is str_word includes dict_word
and it's not identical, if so it pushes the match into the array.
Edited variables names for consistency.